[ppp] Update to 2.4.9 version, reorganize patches. Fixes JB#50020
Update to 2.4.9 which contains many fixes including the fix to
CVE-2020-8597 (commit id 8d7970b8f3db727fe798b65f3377fe6787575426).
Fedora has dropped one of the patches we've been using (see
https://src.fedoraproject.org/rpms/ppp/c/23f594e00896a75b4cf3690aa4a32e0d11917cf7?branch=rawhide )
as that has been upstreamed. We can drop it as well. Modified the
aarch64 build fixing patch (took from Fedora).

Added new patch to revert an IPv6 peer address adding change since the
socket option used (NETLINK_CAP_ACK) is available only with kernel
versions >= 4.3. IPv6 support on the pppd enabled plugins in ConnMan is
not yet implemented so this is kind of "safe" revert as no functionality
is impaired as of not.

Added libpcap-devel as runtime build dep as pppd now depends on it.

Signed-off-by: Jussi Laakkonen <jussi.laakkonen@jolla.com>