[ppp] Update to 2.5.2. Fixes JB#61294
Update to latest to fix CVE-2022-4603 (fix in 2.5.1) and
CVE-2024-58250 (fix in 2.5.2).

Drop the unnecessary patches, there are no Makefile.linux files
anymore and compatibility with older kernels has been fixed.

Exclude also all the example files, we do not ship those.

Add the .pc for pkgconfig in devel package.